Integrate font-awesome
Wow, that was tricky. I was using an old version of grunt-usemin, which
did not recognizes referenced files that have querystrings or hashes.
The newer versions do this. Therefore we bump the version of
grunt-usemin to 2.3.0.

We also now serve fonts from /fonts instead of /styles/fonts, including
vendor fonts (such as those of font-awesome or the glyphicons that come
with bootstrap). This is easier to handle in the Gruntfile, since we can
then easily compile various vendor stylesheets into a single build
asset.

closes #3