Implement calendar using the fruux jCal API
The idea is that we just query the calendar from Fruux.

This way we even get rid of the PHP layer. Requests are proxied through
webserver. Fruux requires the Host header to be set appropriately and
refuses to answer otherwise.

connect-grunt-proxy@0.1.11 unfortunately breaks proxying support, which
is why we moved back to 0.1.10.