Reword verein/spenden
My take on raumzeitlabor/rzl-homepage#6, based on the draft written by
@tabascoeye. Please review punctuation ;)

I am not too happy with the list of examples though, as it stresses donations
from organizations and foundations. That's why I added the last point to the
list. All in all I think it might be better to drop the list completely.