Let Travis do the deployment of the homepage
This commit adds an after_install Travis hook. The hook first ensures it
is running within a Travis environment, decrypts the ssh deployment key
and initializes an ssh-agent, and then copies all files from dist to
premium.

The deployment script is based on the following gist:

  https://gist.github.com/pghalliday/240fe740d523dad21d3f

The ssh deployment key is stored encrypted in the repository (AES-256).
More information on this can be found here:

  http://docs.travis-ci.com/user/encrypting-files/