Fetch git version on request only
Our first approach for #51 always modified the jekyll template to
contain the latest git commit SHA. The disadvantage is that we always
modify all files, even if only a subset actually changed. This means
Travis always deploys all affected HTML files with each build.

This commit instead writes the version information into a separate file
called CHANGELOG.json that will be fetched on demand by the browser as
soon as the user views the footer. This reduces the files required to
be updated on premium and is very conservative in terms of bandwidth.

I was initially afraid to break /bilder where we are also using a
waypoint on the footer. It does however seem to work fine.

closes #51