Revert "Let git generate commit SHA-1 ID on checkout"
This reverts commit 98a406db53c6e06cebcd185056d89de7d2e1b93c.

RTFM: It turns out ident returns the identity of the object, not of
HEAD. The version therefore always stayed the same (since CHANGELOG.json
did not change anymore).

We therefore go back to the previous solution using grunt-git-describe,
but this time with a custom format template. Let's see if this also
works on Travis.