/videos: Switch to new Youtube v3 API
The old API was deprecated on 20th of April, see [1]. We access the new
API using a browser token. This token is whitelisted for the following
referers:

    *.raumzeitlabor.de/*
    *.raumzeitlabor.org/*
    localhost/*

The token is owned by me and can easily be replaced by someone else's
token.

At the same time, we updated the CSP for the new central Google API
endpoint on github.com/raumzeitlabor/rzl-homepage-docker [2].

closes #137

[1] https://support.google.com/youtube/answer/6098135?p=yt_devicesupport&hl=de&rd=1
[2] https://github.com/raumzeitlabor/rzl-homepage-docker/commit/0d5576aa17f73b758e84b3614aa5aac035f2e6cf